+++ murrelet_livecode/src/lib.rs
@@ -4,4 +4,5 @@ pub mod expr;
 pub mod livecode;
 pub mod nestedit;
 pub mod state;
+pub mod types;
 pub mod unitcells;
